Added support for floats

+++ src/interpreter/mod.rs
@@ -3,11 +3,11 @@ use parser::*;
 #[cfg(test)]
 mod tests;
 
-pub fn eval(source: &str) -> i64 {
+pub fn eval(source: &str) -> f64 {
     eval_expr(&Parser::new(source).parse_expression())
 }
 
-fn eval_expr(expr: &Expr) -> i64 {
+fn eval_expr(expr: &Expr) -> f64 {
     match expr {
         Expr::Number(n) => *n,
         Expr::BinExprPtr(box expr) => match expr.op {
